# server/docs/swagger.yaml
basePath: /api/v1
definitions:
  admin.AdminTaskListResponse:
    properties:
      list:
        items:
          $ref: '#/definitions/admin.AdminTaskResponse'
        type: array
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
  admin.AdminTaskResponse:
    properties:
      canForceStop:
        type: boolean
      completedAt:
        type: string
      createdAt:
        type: string
      errorMessage:
        type: string
      id:
        type: integer
      instanceId:
        type: integer
      instanceName:
        type: string
      instanceType:
        type: string
      isForceStoppable:
        type: boolean
      progress:
        type: integer
      providerId:
        type: integer
      providerName:
        type: string
      remainingTime:
        description: 剩余时间（秒）
        type: integer
      startedAt:
        type: string
      status:
        type: string
      statusMessage:
        type: string
      taskType:
        type: string
      timeoutDuration:
        type: integer
      userId:
        type: integer
      userName:
        type: string
      uuid:
        type: string
    type: object
...
  admin.ForceStopTaskRequest:
    properties:
      reason:
        description: 强制停止原因
        type: string
      taskId:
        type: integer
    required:
    - taskId
    type: object
...
  admin.TaskStatsResponse:
    properties:
      completedTasks:
        type: integer
      failedTasks:
        type: integer
      pendingTasks:
        type: integer
      runningTasks:
        type: integer
      timeoutTasks:
        type: integer
      totalTasks:
        type: integer
    type: object
...
  common.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
host: localhost:8888
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: 一键虚拟化管理平台API接口文档
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: OneClickVirt API
  version: "1.0"
paths:
...
  /admin/tasks:
    get:
      consumes:
      - application/json
      description: 获取所有用户的任务列表，支持分页和筛选
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 页大小
        in: query
        name: pageSize
        type: integer
      - description: Provider ID
        in: query
        name: providerId
        type: integer
      - description: 用户ID
        in: query
        name: userId
        type: integer
      - description: 任务类型
        in: query
        name: taskType
        type: string
      - description: 任务状态
        in: query
        name: status
        type: string
      - description: 实例类型
        in: query
        name: instanceType
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.AdminTaskListResponse'
              type: object
        "401":
          description: 权限不足
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取管理员任务列表
      tags:
      - 管理员管理
  /admin/tasks/{taskId}/cancel:
    post:
      consumes:
      - application/json
      description: 管理员取消指定用户的任务
      parameters:
      - description: 任务ID
        in: path
        name: taskId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: 权限不足
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 操作失败
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 管理员取消用户任务
      tags:
      - 管理员管理
  /admin/tasks/force-stop:
    post:
      consumes:
      - application/json
      description: 管理员强制停止运行中的任务
      parameters:
      - description: 强制停止任务请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/admin.ForceStopTaskRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 操作成功
          schema:
            $ref: '#/definitions/common.Response'
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/common.Response'
        "401":
          description: 权限不足
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 操作失败
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 强制停止任务
      tags:
      - 管理员管理
  /admin/tasks/stats:
    get:
      consumes:
      - application/json
      description: 获取系统任务统计信息
      produces:
      - application/json
      responses:
        "200":
          description: 获取成功
          schema:
            allOf:
            - $ref: '#/definitions/common.Response'
            - properties:
                data:
                  $ref: '#/definitions/admin.TaskStatsResponse'
              type: object
        "401":
          description: 权限不足
          schema:
            $ref: '#/definitions/common.Response'
        "500":
          description: 获取失败
          schema:
            $ref: '#/definitions/common.Response'
      security:
      - BearerAuth: []
      summary: 获取任务统计
      tags:
      - 管理员管理
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"